Handle multiple top states
- also fix site_tool build (used to find artists with multiple states as #1).
