+++ Sources/App/Localizable.xcstrings
@@ -333,6 +333,9 @@
     "Multiple Months with Most Shows" : {
 
     },
+    "Multiple States with Most Shows" : {
+
+    },
     "Multiple Weekdays with Most Shows" : {
 
     },
